// src/payload-types.ts
export interface Article {
  id: string
  title: string
  slug: string
  status: 'draft' | 'published'
  updatedAt: string
  createdAt: string
}